# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UVA Production Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new client account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated

  # ── Contact & newsletter ─────────────────────────────────────
...
  /meridian/disputes/{id}:
    patch:
      operationId: updateMeridianDispute
      tags: [meridian]
      summary: Update a dispute status/priority and optionally append a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeUpdateInput"
      responses:
        "200":
          description: Updated dispute
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderDispute"

  # ── Client portal ────────────────────────────────────────────
...
  /cadence/engagements/{id}:
    get:
      operationId: getCadenceEngagement
      tags: [cadence]
      summary: Get one engagement with its milestones and deliverables
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Engagement detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CadenceEngagementDetail"
        "403":
          description: Not your engagement
        "404":
          description: Not found
...
  /cadence/milestones/{id}:
    patch:
      operationId: updateCadenceMilestone
      tags: [cadence]
      summary: Update a milestone status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CadenceMilestone"
        "403":
          description: Not your milestone
        "404":
          description: Not found

# ── Component Schemas ───────────────────────────────────────────────
components:
  schemas:
...
    ProductKey:
      type: string
      description: The product/portal a user belongs to.
      enum:
        [hub, cadence, rise, compass, meridian, spark, aria, pulse, sentinel, tend]
    AuthUser:
      type: object
      required: [id, email, name, role, productKey, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        organization:
          type: ["string", "null"]
        role:
          type: string
        productKey:
          $ref: "#/components/schemas/ProductKey"
        createdAt:
          type: string
          format: date-time
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        organization:
          type: string
        productKey:
          $ref: "#/components/schemas/ProductKey"
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
    ProviderDisputeNote:
      type: object
      required: [author, body, at]
      properties:
        author:
          type: string
        body:
          type: string
        at:
          type: string
          format: date-time
    ProviderDispute:
      type: object
      required: [id, subject, category, status, priority, notes, createdAt, updatedAt]
      properties:
        id:
          type: integer
        providerId:
          type: ["integer", "null"]
        subject:
          type: string
        category:
          type: string
        status:
          type: string
        priority:
          type: string
        notes:
          type: array
          items:
            $ref: "#/components/schemas/ProviderDisputeNote"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DisputeUpdateInput:
      type: object
      properties:
        status:
          type: string
          enum: [Open, In review, Resolved, Escalated]
        priority:
          type: string
          enum: [Low, Normal, High, Urgent]
        note:
          type: string

    # ── Client portal ─────────────────────────────────────────────
...
    CadenceEngagement:
      type: object
      required:
        [
          id,
          userId,
          title,
          practiceArea,
          status,
          createdAt,
          milestoneCount,
          deliverableCount,
          completedDeliverableCount,
        ]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        practiceArea:
          type: string
        status:
          type: string
        nextMilestone:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        milestoneCount:
          type: integer
        deliverableCount:
          type: integer
        completedDeliverableCount:
          type: integer
    CadenceMilestone:
      type: object
      required: [id, engagementId, title, status, orderIndex, createdAt]
      properties:
        id:
          type: integer
        engagementId:
          type: integer
        title:
          type: string
        status:
          type: string
        dueDate:
          type: ["string", "null"]
          format: date-time
        orderIndex:
          type: integer
        createdAt:
          type: string
          format: date-time
    CadenceDeliverable:
      type: object
      required: [id, engagementId, title, status, qaGateStatus, createdAt, updatedAt]
      properties:
        id:
          type: integer
        engagementId:
          type: integer
        milestoneId:
          type: ["integer", "null"]
        title:
          type: string
        status:
          type: string
        qaGateStatus:
          type: string
          description: pending, passed, or failed
        qaNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CadenceEngagementDetail:
      type: object
      required: [engagement, milestones, deliverables]
      properties:
        engagement:
          $ref: "#/components/schemas/CadenceEngagement"
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/CadenceMilestone"
        deliverables:
          type: array
          items:
            $ref: "#/components/schemas/CadenceDeliverable"
...
    MilestoneUpdateInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
...