# lib/api-zod/src/generated/types/cadenceEngagement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UVA Production Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface CadenceEngagement {
  id: number;
  userId: number;
  title: string;
  practiceArea: string;
  status: string;
  /** @nullable */
  nextMilestone?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dueDate?: Date | null;
  createdAt: Date;
  milestoneCount: number;
  deliverableCount: number;
  completedDeliverableCount: number;
}