# lib/api-zod/src/generated/types/cadenceEngagementDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UVA Production Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { CadenceDeliverable } from './cadenceDeliverable';
import type { CadenceEngagement } from './cadenceEngagement';
import type { CadenceMilestone } from './cadenceMilestone';

export interface CadenceEngagementDetail {
  engagement: CadenceEngagement;
  milestones: CadenceMilestone[];
  deliverables: CadenceDeliverable[];
}