# lib/api-zod/src/generated/types/productKey.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UVA Production Engine API
 * OpenAPI spec version: 0.1.0
 */

/**
 * The product/portal a user belongs to.
 */
export type ProductKey = typeof ProductKey[keyof typeof ProductKey];


export const ProductKey = {
  hub: 'hub',
  cadence: 'cadence',
  rise: 'rise',
  compass: 'compass',
  meridian: 'meridian',
  spark: 'spark',
  aria: 'aria',
  pulse: 'pulse',
  sentinel: 'sentinel',
  tend: 'tend',
} as const;